rename generator and output
