+++ bindings/generate_bindings.py
